requirements.txt
deepdish==0.3.7
gpustat==0.6.0
h5py==2.10.0
numpy==1.19.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
torchkeras==2.1.2